@@ -26,7 +26,7 @@
 		<dependency>
 			<groupId>org.mockito</groupId>
 			<artifactId>mockito-all</artifactId>
-			<version>1.9.5</version>
+			<version>1.10.19</version>
 		</dependency>
 	</dependencies>
 
@@ -51,8 +51,8 @@
 		<artifactId>maven-compiler-plugin</artifactId>
 		<version>3.5.1</version>
 		<configuration>
-			<source>8</source>
-			<target>8</target>
+			<source>17</source>
+			<target>17</target>
 		</configuration>
 	</plugin>
 	</plugins>
