@@ -2,37 +2,38 @@ package repository;
 
 import database.DBManager;
 import model.Item;
-import org.junit.BeforeClass;
-import org.junit.AfterClass;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.notNullValue;
-import static org.junit.Assert.assertThat;
 import static org.hamcrest.core.Is.is;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 /**
  * @author evrnsky <vrnsky at protonmail.ch>
  * @version 0.1
  * @since 11.04.2017
- *
+ * <p>
  * Unit test for item repository.
  */
-public class ItemRepositoryTest {
+class ItemRepositoryTest {
 
     /**
      * Before each test case we need to build session factory.
      */
-    @BeforeClass
+    @BeforeAll
     public static void setUp() {
         DBManager.getInstance().buildSessionFactory();
     }
 
     /**
-     * When try add item should check that item was added.
+     * When try to add item should check that item was added.
      */
     @Test
-    public void whenTryAddItemShouldCheckThatItemWasAdded() {
+    void whenTryAddItemShouldCheckThatItemWasAdded() {
         Item item = new Item("Value");
         ItemRepository.getInstance().addItem(item);
         assertThat(item.getId() != 0, is(true));
@@ -43,7 +44,7 @@ public class ItemRepositoryTest {
      * When try edit item should check that data saved.
      */
     @Test
-    public void whenTryEditItemShouldCheckThatDataSaved() {
+    void whenTryEditItemShouldCheckThatDataSaved() {
         Item item = new Item("value");
         ItemRepository.getInstance().addItem(item);
         item.setDesc("new value");
@@ -52,10 +53,10 @@ public class ItemRepositoryTest {
     }
 
     /**
-     * When try remove item should check that item removed.
+     * When try to remove item should check that item removed.
      */
     @Test
-    public void whenTryRemoveItemShouldCheckThatItemRemoved() {
+    void whenTryRemoveItemShouldCheckThatItemRemoved() {
         Item item = new Item("value");
         ItemRepository.getInstance().addItem(item);
         ItemRepository.getInstance().addItem(item);
@@ -67,7 +68,7 @@ public class ItemRepositoryTest {
      * When try to show only done task should check that correct.
      */
     @Test
-    public void whenTryToShowOnlyAlreadyDoneTaskShouldCheckThatAllIsOk() {
+    void whenTryToShowOnlyAlreadyDoneTaskShouldCheckThatAllIsOk() {
         Item item = new Item("not done");
         item.setDone(false);
         Item twoItem = new Item("done");
@@ -79,10 +80,10 @@ public class ItemRepositoryTest {
     }
 
     /**
-     * When try get item by id should check that all is ok.
+     * When try to get item by id should check that all is ok.
      */
     @Test
-    public void whenTryGetItemByIdShouldCheckThatIsCorrectItem() {
+    void whenTryGetItemByIdShouldCheckThatIsCorrectItem() {
         Item item = new Item("value");
         ItemRepository.getInstance().addItem(item);
         assertEquals(ItemRepository.getInstance().getById(item.getId()).getDesc(), "value");
@@ -91,8 +92,8 @@ public class ItemRepositoryTest {
     /**
      * After each test close session factory.
      */
-    @AfterClass
-    public static   void tearDown() {
+    @AfterAll
+    public static void tearDown() {
         DBManager.getInstance().closeSessionFactory();
     }
 }
