@@ -2,13 +2,15 @@ package controllers;
 
 import dao.Repository;
 import models.User;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
-import static org.junit.Assert.assertEquals;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -16,19 +18,20 @@ import static org.mockito.Mockito.when;
  * @author evrnsky
  * @version 0.1
  * @since 05.03.2017
- *
+ * <p>
  * This unit test check work of adding new user servlet.
  */
-public class CreateUserTest {
+class CreateUserTest {
 
 
     /**
      * When admin add new user to the system should check that is ok.
+     *
      * @throws ServletException if problem with concurrency.
-     * @throws IOException if problem with data exchange.
+     * @throws IOException      if problem with data exchange.
      */
     @Test
-    public void whenAdminAddNewUserShouldCheckThatUserWasAdded() throws ServletException, IOException {
+    void whenAdminAddNewUserShouldCheckThatUserWasAdded() throws ServletException, IOException {
         CreateUser createUser = new CreateUser();
         HttpServletRequest request = mock(HttpServletRequest.class);
         HttpServletResponse response = mock(HttpServletResponse.class);
@@ -38,6 +41,6 @@ public class CreateUserTest {
         createUser.doPost(request, response);
         List<User> users = Repository.getInstance().getAllUsers();
         User added = users.get(users.size() - 1);
-        assertEquals(added.getFirstName(), "Andrew");
+        assertEquals("Andrew", added.getFirstName());
     }
 }
