@@ -2,14 +2,15 @@ package controllers;
 
 import database.Repository;
 import models.User;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
-import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -17,18 +18,19 @@ import static org.mockito.Mockito.when;
  * @author evrnsky
  * @version 0.1
  * @since 05.03.2017
- *
+ * <p>
  * This unit test check work of edit controller of app.
  */
-public class EditUserTest {
+class EditUserTest {
 
     /**
      * Add new user to the system and edit it, after it operations check that all is ok.
+     *
      * @throws ServletException if problem with concurrency.
-     * @throws IOException if problem with data exchange.
+     * @throws IOException      if problem with data exchange.
      */
     @Test
-    public void whenAdminEditUserShouldCheckThatDataWasChanged() throws ServletException, IOException {
+    void whenAdminEditUserShouldCheckThatDataWasChanged() throws ServletException, IOException {
         CreateUser createUser = new CreateUser();
         HttpServletRequest request = mock(HttpServletRequest.class);
         HttpServletResponse response = mock(HttpServletResponse.class);
