@@ -2,14 +2,16 @@ package controllers;
 
 import database.Repository;
 import models.User;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsNull.nullValue;
-import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -20,7 +22,7 @@ import static org.mockito.Mockito.when;
  *
  * This unit test check controller's work for removing user from app.
  */
-public class RemoveUserTest {
+class RemoveUserTest {
 
 
     /**
@@ -29,7 +31,7 @@ public class RemoveUserTest {
      * @throws IOException if problem with data exchange.
      */
     @Test
-    public void whenAdminRemoveUserShouldCheckThatUserWasRemoved() throws ServletException, IOException {
+    void whenAdminRemoveUserShouldCheckThatUserWasRemoved() throws ServletException, IOException {
         CreateUser createUser = new CreateUser();
         HttpServletRequest request = mock(HttpServletRequest.class);
         HttpServletResponse response = mock(HttpServletResponse.class);
