@@ -1,14 +1,15 @@
 package database;
 
 import models.User;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsNull.nullValue;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 /**
  * @author evrnsky
@@ -17,7 +18,7 @@ import static org.junit.Assert.assertEquals;
  *
  * This is unit test for repository.
  */
-public class RepositoryTest {
+class RepositoryTest {
 
     /**
      * Instance of system under test.
@@ -28,17 +29,16 @@ public class RepositoryTest {
      * This method calls before each test case.
      * It instant repository instance.
      */
-    @Before
+    @BeforeEach
     public void setUp() {
         repo = Repository.getInstance();
     }
 
     /**
-     * When try add user should check that user was added.
-     * @throws Exception if something wrong.
+     * When try to add user should check that user was added.
      */
     @Test
-    public void whenTryAddUserShouldCheckThatUserWasAdded() throws Exception {
+    void whenTryAddUserShouldCheckThatUserWasAdded()  {
         User user = new User("Andrew", "Voronyansky", "vrnsky@vrnsky.com");
         this.repo.addUser(user);
         assertThat(this.repo.findUserById(user.getId()), is(user));
@@ -46,10 +46,9 @@ public class RepositoryTest {
 
     /**
      * When try edit user should check that data is saved.
-     * @throws Exception if something wrong.
      */
     @Test
-    public void whenTryEditUserShouldCheckThatRepoAcceptedNewVersion() throws Exception {
+    void whenTryEditUserShouldCheckThatRepoAcceptedNewVersion() {
         User user = new User("Andrew", "Voronyansky", "vrnsky@vrnsky.com");
         this.repo.addUser(user);
         user.setEmail("wolf@vrn.com");
@@ -59,10 +58,9 @@ public class RepositoryTest {
 
     /**
      * When remove user should check that method find by id return null.
-     * @throws Exception if some error happened.
      */
     @Test
-    public void whenRemoveUserShouldCheckThatMethodFindByIdReturnNull() throws Exception {
+    void whenRemoveUserShouldCheckThatMethodFindByIdReturnNull() {
         User user = new User("yegor", "256", "eo");
         this.repo.addUser(user);
         this.repo.removeUser(user);
@@ -70,11 +68,10 @@ public class RepositoryTest {
     }
 
     /**
-     * When try get all users should check that method return correct list.
-     * @throws Exception if some error happened.
+     * When try to get all users should check that method return correct list.
      */
     @Test
-    public void whenTryGetAllUsersShouldCheckThatMethodReturnCorrectData() throws Exception {
+    void whenTryGetAllUsersShouldCheckThatMethodReturnCorrectData() {
         User user = new User("yegor", "256", "eo");
         this.repo.addUser(user);
         assertEquals(1, this.repo.getAllUsers().size());
@@ -84,7 +81,7 @@ public class RepositoryTest {
      * After all test we need to clear storage, because it is thread safe storage.
      * And without invoke clear method we have a problem with size test cases.
      */
-    @After
+    @AfterEach
     public void tearDown() {
         this.repo.clear();
     }
