@@ -3,6 +3,6 @@
  * @version 0.1
  * @since 01.04.2017
  *
- * This package contains test for user interface.
+ * This package contains a unit test for model package.
  */
-package start;
+package ru.evrnsky.model;
